% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawr.R
\name{drawr}
\alias{drawr}
\title{Drawr}
\usage{
drawr(data, x_col, y_col, draw_start, title = NULL, pin_start = TRUE,
  x_range = NULL, y_range = NULL, line_style = NULL,
  drawn_line_color = "orangered", data_line_color = "steelblue",
  x_axis_buffer = 0.02, y_axis_buffer = 0.1)
}
\arguments{
\item{data}{Tibble to draw}

\item{x_col}{Name of column for x axis}

\item{y_col}{Name of column for y axis}

\item{draw_start}{Where on the x axis to start obscuring data for drawing?}

\item{title}{Text for title, if desired.}

\item{pin_start}{Pin start of drawn line to end of shown data? Defaults to `TRUE`.}

\item{x_range}{Two element array of min and max of x range. Otherwise defaults to min and max of data.}

\item{y_range}{Two element array of min and max of y range. Otherwise defaults to min and max of data.}

\item{line_style}{List containing any styling that is desired for the default line. For options see [MDN SVG line](https://developer.mozilla.org/en-US/docs/Web/SVG/Element/line). Make sure to camelCase all attributes.}

\item{drawn_line_color}{CSS valid color for user-drawn line. Defaults to `"orangered"`.}

\item{data_line_color}{CSS valid color for data line. Defaults to `"steelblue"`.}

\item{x_axis_buffer}{Scaler for how much to pad ends of x axis if not specifying range directly. Defaults to 2% (`0.02`).}

\item{y_axis_buffer}{Scaler for how much to pad ends of y axis if not specifying range directly. Defaults to 10% (`0.1`).}
}
\value{
Interactive you-draw-it plot.
}
\description{
Drawr
}
\examples{
drawr(data = tibble(x = 1:50, y = sin(x)), x_col = x, y_col = y, draw_start = 25)
}
