% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawr_widget.R
\name{drawr_widget}
\alias{drawr_widget}
\title{Creates and embeds a nice you draw it chart}
\usage{
drawr_widget(data, draw_start, raw_draw = FALSE, x_key = "x",
  y_key = "y", y_min = NA, y_max = NA, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{dataframe containing data you want to plot in two of its columns}

\item{draw_start}{position on the x-axis the true data is blocked off and the user is to draw from.}

\item{raw_draw}{set to true if you want to not draw any line, just let the user draw everything. Auto sets draw_start to begining of data.}

\item{x_key}{name of the x column.}

\item{y_key}{name of the y column.}

\item{y_min}{value of the lowest possible value the user is allowed to draw, defaults to lowest seen in data.}

\item{y_max}{value of the highest possible value the user is allowed to draw, defaults to highest seen in data.}
}
\description{
Makes a draw chart that can go into an rmarkdown or a shiny app.
}
