% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{run_demo}
\alias{run_demo}
\title{Run demo apps
Load and play with simple demo apps using the different senses available in shinysense. Currently supports
`draw_distribution` which allows the user to draw a beta distribution and get its parameters.}
\usage{
run_demo(demo_name)
}
\arguments{
\item{demo_name}{Name of demo app to run. Options `{'draw_distribution'}`.}
}
\value{
Shiny app instance
}
\description{
Run demo apps
Load and play with simple demo apps using the different senses available in shinysense. Currently supports
`draw_distribution` which allows the user to draw a beta distribution and get its parameters.
}
\examples{
run_demo('draw_distribution')
}
