% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydrawr.R
\name{shinydrawr_UI}
\alias{shinydrawr_UI}
\title{Display a lineplot that lets user's draw data: Server version}
\usage{
shinydrawr_UI(id, height = "400px")
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}

\item{height}{How tall the plot should be in css valid units. Defaults to
`'400px'`.}
}
\value{
UI portion of the shinydrawr module
}
\description{
Make a you draw it line chart in the style of this [New York Times
article](https://www.nytimes.com/interactive/2017/01/15/us/politics/you-draw-obama-legacy.html?mtrref=undefined&gwh=D06000A9C788821324D9EED3BCA9C3D1&gwt=pay)
about  button for UI. Wherever this is placed in the UI it will make a line
chart that can be drawn on.
}
\examples{
if(interactive()){
shinydrawr_UI("drawr_widget", height = '300px')
}
}
\seealso{
\code{\link{shinydrawr}}, \code{\link{drawr}}
}
