% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydrawr.R
\name{shinydrawrUI}
\alias{shinydrawrUI}
\title{Make a you draw it line chart in the style of this [New York Times article](https://www.nytimes.com/interactive/2017/01/15/us/politics/you-draw-obama-legacy.html?mtrref=undefined&gwh=D06000A9C788821324D9EED3BCA9C3D1&gwt=pay) about  button for UI.
   Wherever this is placed in the UI it will make a line chart that can be drawn on.}
\usage{
shinydrawrUI(id)
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}

\item{...}{additional tags to \code{\link{div}} to display before plot is
displayed}
}
\value{
A blue button that you press to initiate or stop recording of sound.
}
\description{
Make a you draw it line chart in the style of this [New York Times article](https://www.nytimes.com/interactive/2017/01/15/us/politics/you-draw-obama-legacy.html?mtrref=undefined&gwh=D06000A9C788821324D9EED3BCA9C3D1&gwt=pay) about  button for UI.
   Wherever this is placed in the UI it will make a line chart that can be drawn on.
}
\examples{
shinydrawrUI('myrecorder')
}
