% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyearr.R
\name{shinyearr}
\alias{shinyearr}
\title{Gather recorded data from UI.}
\usage{
shinyearr(input, output, session)
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}
}
\description{
Exports a reactive array of length 256, corresponding to a fourier transform of the sound waves of your recoding. This is a frequently used format for running various speech recognition algorithms on. Future edditions will allow access to the raw data.
    This is the server component of shinyearr. You never directly use this function but instead call it through the shiny function `callModule()`. See the example for how to do this.
}
\examples{
\dontrun{
shiny::callModule(shinyearr, "myrecorder")
}
}
