% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyearr.R
\name{shinyearrUI}
\alias{shinyearrUI}
\title{Make a record button for UI.}
\usage{
shinyearrUI(id)
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}
}
\value{
A blue button that you press to initiate or stop recording of sound.
}
\description{
Make a record button for UI.
}
\examples{
shinyearrUI('myrecorder')
}
