% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylistenr.R
\name{shinylistenr_UI}
\alias{shinylistenr_UI}
\title{Access data from user's microphone: UI version}
\usage{
shinylistenr_UI(id)
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}
}
\value{
A blue button that you press to initiate or stop recording of sound.
}
\description{
Access data from user's microphone: UI version
}
\examples{
\dontrun{
shinylistenr_UI("recorder")
}
}
\seealso{
\code{\link{shinylistenr}}
}
