% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymovr.R
\name{shinymovr}
\alias{shinymovr}
\title{Gather recorded data from UI.}
\usage{
shinymovr(input, output, session)
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}
}
\description{
Contains a reactive variable that will update as new data comes in.
}
\examples{
\dontrun{
movrData <- callModule(shinymovr)
}
}
