% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymovr.R
\name{shinymovr_UI}
\alias{shinymovr_UI}
\title{Access data from user's accelerometers: UI function}
\usage{
shinymovr_UI(id, resting_msg = "Turn On", button_width = "130px",
  button_height = "50px", button_color = "steelblue")
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}

\item{resting_msg}{the string displayed when the button is not pressed.
Defaults to "Turn On"}

\item{button_width}{width of the button in a valid css string. E.g. needs to
have pixels appended to it.}

\item{button_height}{height of the button, again in valid css string.}

\item{button_color}{valid css string to control the button color.}
}
\value{
A blue button that you press to initiate or stop recording of
  acceleration data.
}
\description{
Use shiny on your mobile phone to gather accelerometer data for whatever your
heart could desire. Note that this has been recently hamstrung by browser's
attempts to better secure user data. It still works on mobile devices but
motion detection needs to be enabled in the devices settings.
}
\examples{
\dontrun{
shinymovr_UI(
  'movr_button',
  resting_msg = 'Click me to record',
  button_width = '200px')
}
}
\seealso{
\code{\link{shinymovr}}
}
