% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymovr.R
\name{shinymovrUI}
\alias{shinymovrUI}
\title{Use shiny on your mobile phone to gather accelorameter data for whatever your heart could desire.}
\usage{
shinymovrUI(id)
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}
}
\value{
A blue button that you press to initiate or stop recording of acceloration data.
}
\description{
Use shiny on your mobile phone to gather accelorameter data for whatever your heart could desire.
}
\examples{
shinymovrUI('movrButton')
}
