% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypopup.R
\name{shinypopup}
\alias{shinypopup}
\title{Listens for clicking of accept button and kills popup if clicked.}
\usage{
shinypopup(input, output, session)
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}
}
\description{
Just sits there, no need to pass it anything.
}
\examples{
callModule(shinypopup, "myTerms")
}
