% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypopup.R
\name{shinypopupUI}
\alias{shinypopupUI}
\title{Make a popup button that users have to accept before seeing content.}
\usage{
shinypopupUI(id, buttonText, popupDiv, ...)
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}

\item{buttonText}{text that appears in the button for accepting the terms of the popup}

\item{popupDiv}{standard html content for the popup card. E.g. `div("my terms are")`}

\item{...}{additional argument to pass to \code{\link{tagList}}}
}
\value{
A popup that blurs everything behind it and dissapears upon a button click.
}
\description{
Make a popup button that users have to accept before seeing content.
}
\examples{
\dontrun{
shinypopupUI('myTerms')
}
}
