% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyswipr.R
\name{shinySwipr}
\alias{shinySwipr}
\title{shinyswipr server function}
\usage{
shinySwipr(input, output, session)
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}
}
\description{
This is the server component of the shiny swipr app.
You never call this function directly but instead call it
from the `callModule` syntax that shiny uses for modules.
See the example for how to do this.
}
\examples{
callModule(shinySwipr, "myswipr")
}
