% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyswipr.R
\name{shinyswipr_UI}
\alias{shinyswipr_UI}
\title{Add a swipable card component to shiny app: UI function}
\usage{
shinyswipr_UI(id, ...)
}
\arguments{
\item{id}{The id of the current element you are entering}

\item{...}{Any other UI elements you wish to include within the swiping zone.}
}
\description{
Wraps whatever UI elements are passed to it in a card format that can then be
swiped to the left, right, up, or down with the results of that action being
sent to the shiny app.
}
\examples{
\dontrun{
shinyswipr_UI(id = "myswipr", h1("This is my title"), p("here is some text"))
}
}
\seealso{
\code{\link{shinyswipr}}
}
