% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyviewr.R
\name{shinyviewrUI}
\alias{shinyviewrUI}
\title{A webcam view and snapshot function. Will send a 3d array (width, height, colors+opacity) back to your computer of the frame when you click take photo.}
\usage{
shinyviewrUI(id)
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}

\item{...}{additional tags to \code{\link{div}} to display before plot is
displayed}
}
\value{
A video overlay and a 'shutter' button.
}
\description{
A webcam view and snapshot function. Will send a 3d array (width, height, colors+opacity) back to your computer of the frame when you click take photo.
}
\examples{
\dontrun{
shinyviewrUI('myrecorder')
}
}
