% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewr_widget.R
\name{viewr_widget-shiny}
\alias{viewr_widget-shiny}
\alias{viewr_widgetOutput}
\alias{renderViewr_widget}
\title{Shiny bindings for viewr_widget}
\usage{
viewr_widgetOutput(outputId, width = "100\%", height = "400px")

renderViewr_widget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a viewr_widget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using viewr_widget within Shiny
applications and interactive Rmd documents.
}
