% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewr_widget.R
\name{viewr_widget}
\alias{viewr_widget}
\title{Widget form of viewr.}
\usage{
viewr_widget(outputWidth = NULL, outputHeight = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{outputWidth}{The number of pixels in width desired from the output frame. Defaults to the size of the container}

\item{outputHeight}{The number of pixels in height desired from the output frame. Defaults to size of container.}

\item{width}{How many pixels wide the embeded widget should be (usually left as \code{NULL} to be automatically decided by the function.)}

\item{height}{How many pixels high the embedded widget should be (again, usually left as \code{NULL}).}

\item{elementID}{Explicit ID of the widget. Like \code{width} and \code{height} this is usually left as \code{NULL} for R to automatically Assign.}
}
\description{
This is an internal function for use in the function \code{shinviewr}. It is an htmlwidget that loads a webcam feed to the page. Getting images back from it requires some manipulation that \code{shinyviewr} takes care of.
}
