% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{run_demo}
\alias{run_demo}
\title{Run demo apps
Load and play with simple demo apps using the different senses available in shinysense. Currently supports
`draw_distribution` which allows the user to draw a beta distribution and get its parameters.`swipr` which demos basic
usage of cards and decision recording with `shinyswipr`.}
\usage{
run_demo(demo_name, show_code = TRUE)
}
\arguments{
\item{demo_name}{Name of demo app to run. Options `{'shinydrawr', 'shinyviewr', 'shinyswipr'}`.}

\item{show_code}{See the underlying code for the demo app? Defaults to `TRUE`./}
}
\value{
Shiny app instance
}
\description{
Run demo apps
Load and play with simple demo apps using the different senses available in shinysense. Currently supports
`draw_distribution` which allows the user to draw a beta distribution and get its parameters.`swipr` which demos basic
usage of cards and decision recording with `shinyswipr`.
}
\examples{
run_demo('shinydrawr')
}
