% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylistenr.R
\name{shinylistenr_UI}
\alias{shinylistenr_UI}
\title{Make a record button for UI.}
\usage{
shinylistenr_UI(id)
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}
}
\value{
A blue button that you press to initiate or stop recording of sound.
}
\description{
Make a record button for UI.
}
\examples{
shinylistenr_UI('myrecorder')
}
