% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypopup.R
\name{shinypopup}
\alias{shinypopup}
\title{Listens for clicking of accept button and kills popup if clicked.}
\usage{
shinypopup(input, output, session, accepted = FALSE)
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}

\item{accepted}{Logical variable indicating if the user has accepted the terms already. Useful if the page resets for a login. Defaults to FALSE.}
}
\description{
Just sits there, no need to pass it anything.
}
\examples{
\dontrun{
shiny::callModule(shinypopup, "myTerms", accepted = FALSE)
}
}
