% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyswipr.R
\name{shinyswipr}
\alias{shinyswipr}
\title{Add a swipeable card to shiny app: server function}
\usage{
shinyswipr(input, output, session)
}
\arguments{
\item{input}{you can ignore these as it is taken care of by
\code{shiny::callModule()}}

\item{output}{you can ignore these as it is taken care of by
\code{shiny::callModule()}}

\item{session}{you can ignore these as it is taken care of by
\code{shiny::callModule()}}
}
\description{
This is the server component of the shiny swipr app. You never directly use
this function but instead call it through the shiny function `callModule()`.
See the example for how to do this.
}
\examples{
\dontrun{
callModule(shinyswipr, "myswipr")
}
}
\seealso{
\code{\link{shinyswipr_UI}}
}
