% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyswipr.R
\name{shinyswiprUI}
\alias{shinyswiprUI}
\title{shinyswipr UI function}
\usage{
shinyswiprUI(id, ...)
}
\arguments{
\item{id}{The id of the current element you are entering}

\item{...}{Any other UI elements you wish to include within the swiping zone.}
}
\description{
This is the UI component of the shiny swipr app
}
\examples{
\dontrun{
shinyswiprUI(id = "myswipr", h1("This is my title"), p("here is some text"))
}
}
