% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyviewr.R
\name{shinyviewr}
\alias{shinyviewr}
\title{Gather recorded data from UI.}
\usage{
shinyviewr(input, output, session, outputWidth = NULL, outputHeight = NULL)
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}

\item{outputWidth}{How many pixels wide you want your returned photos. Defaults to 200px}

\item{outputHeight}{How many pixels hight you want your returned photos. Defaults to 150px}
}
\description{
Upon completion of line draw, returns a reactive variable that contains a vector of the y coordinates of what the user has drawn. This also includes the start point specified with `draw_start`
    This is the server component of shinyviewr. You never directly use this function but instead call it through the shiny function `callModule()`. See the example for how to do this.
}
\examples{
\dontrun{
 drawChart <- shiny::callModule(shinyviewr, "myCamera")
 }
}
